requirements.txt
Flask
gunicorn
azure-identity
azure-mgmt-costmanagement
python-dotenv